+++ lib/pango/function-1.16.go
@@ -25,7 +25,7 @@ import call "github.com/pekim/gobbi/lib/internal/call"
 func Version() int32 {
 	data := call.Data{Return: call.Return{Type: call.RT_INT}}
 	call.Function(9015, &data)
-	ret := int32(data.Return.Int)
+	ret := data.Return.Int32()
 
 	return ret
 }
